struct: Extract TemplateNode class to separate file
